+++ frontend/js/components/userCard/userCard.js
@@ -1,4 +1,4 @@
-import { changeStatus, fetchAuth, loadHTMLComponent } from "../../api/fetchData.js";
+import { fetchAuth, loadHTMLComponent } from "../../api/fetchData.js";
 import { navigateTo } from "../../router.js";
 import { closeAlertAvatar, closeAlertInfo, setupSettings, closeAlert2FA, clearSettings } from "./utils.js";
 import { disable2FA, enable2FA, updateMenu2FA, checkConfirmationCode } from "./menu2FA.js";
